refactor(services): reconfigure services
This changes is needed for extensibility of the BaseService.
BaseService has delegated some of the common methods used in the _util property, which is a class of type UtilService.
Data is now treated as a mock data handled by MockService class.
No more dependencies to axios for service, since we mock the data.