REFACTOR(confidentialities.selector): use selector to get the data from redux state
Before: Confidentialities component is highly coupled with the state shape of the redux state, i.e. from the mapStateToProps function.

After: It is no longer coupled, because the use of the selector. MapStateToProps function now accepts the whole state, instead of the slice of the state using object destructuring. The slicing is fully done by the selectors

REFACTOR(confidentialities.selector): use reselect library to construct the selectors