REFACTOR: create a higher-order action creators, selectors, reducers
Before:
For each resources, we need to create action creators, selectors, and reducers.
This is fine but it seems like in our use case, all those three are duplicated across resources.

After:
By using the factories, we can just use the factories to create common action creators, selectors and reducers.